Moved file into different package. Updated godoc
