+++ state/traverse.go
@@ -1,4 +1,4 @@
-package util
+package state
 
 import (
 	"fmt"
@@ -10,6 +10,11 @@ import (
 // functor that will be get called on each block header when traversing blocks.
 type onBlockTraverse = func(header *flow.Header) (bool, error)
 
+// TraverseBlocksBackwards traverses a chain segment beginning with the start block (inclusive)
+// Blocks are traversed in reverse
+// height order, meaning the end block must be an ancestor of the start block.
+// The callback is called for each block in this segment.
+// Return value of callback is used to decide if it should continue or not.
 func TraverseBlocksBackwards(headers storage.Headers, startBlockID flow.Identifier, traverse onBlockTraverse) error {
 	ancestorID := startBlockID
 	for {
